refactor: useContext를 함께 모아서 제공하는 방식으로 수정
- export 문이 함께 있어도 괜찮은가 아직 의문점이 있음
